strace: if using libaio, make sure it compiles first.
When using parallel builds, we could end up compiling libaio at the same
time as strace, giving weird race conditions.  strace compiles okay with or
without libaio, but not with it half-installed.

The reason we have libaio is that libfio depends on it, although they forgot
to declare that in Config.in, so we update fio/Config.in too.

Change-Id: Iaa006055f2e81cdda582f87baa6bb771cce42a5f