Set bridge priority to 0x7990 for fat bruno.
Currently the system with the lowest mac address becomes the root of
the spanning tree. If the thin bruno has a lower mac address than the
fat bruno, it becomes the root. The problem with this scenario is that,
if another thin bruno is plugged in the network, or one that's the root
is removed, the GE driver cannot handle the change gracefully, resulting
in nondeterministic behavior, such as pauses or even hangs. Giving the
fat bruno a lower bridge priority value will enable it to be the root
of the spanning tree. The default bridge priority is 0x8000.  The value
0x7990 is chosen to allow some flexibility in case we would like devices
other than the fat bruno to become root of the tree, while guaranteeing
that if the only devices plugged in are brunos, the fat bruno will be
the root.

Change-Id: Id1df2220be7fed88e38b0ffb0c8bde346249fedc