dropbear: disable reverse dns and password authentication.
Reverse DNS is not a security feature, but it can cause long delays if your
DNS is broken (like if you're not connected to the Internet).  Turn it off.

Password authentication is very insecure for ssh, but we might still want it
for getty, so we can't remove our root password (yet?).  Disable all
password authentication in dropbear so at least you need a cert to login.

While we're here, remove all the authorized ssh certs except the "official"
lab one.

Change-Id: I43b0524aeded6fc5dfe14dcf02f64c6775c9eb48