sageserver: babysit sageserver/sageclient in case they die.
As with several of our other daemons, let's run these under the babysitter
and auto-restart them within a few seconds of failure.

While I was here, I cleaned up runsageclient to match an earlier cleanup of
runsage.

Moved the 'logger -t s' redirect out of runsage/runsageclient and into
/etc/init.d/, where it is for most of our other services.  This was also
necessary so that runsage/runsageclient would correctly return nonzero if
siege returned nonzero (like if it died unintentionally), so that babysit
would know to restart it.  Otherwise the return code of runsage is the
return code of 'logger', which is silly.

Change-Id: Ie0d02518540bcce292a94b8c078f283b67bc145e