google_keystore_client: try static files first.
This is an updated version of commit
4dc46736ec9215aeeae0f94d1a9a438befa21972 to work again with the new-style
keystore_client stuff.  Each package declares which keys it plans to need.
If all those keys are available from /usr/local/google/gfiber as plain
files, we don't bother doing the 'blaze build' of the keystore client.  That
way an autobuilder can build the image without needing blaze access.

Change-Id: I9a0cb4813f3b5135b15393d5dd208aab7e38c86e