Sage.properties: Add rc6 codes for Philips S 0099.
This is from a random code in a universal remote I have.  With these codes,
setting the universal remote to Philips S 0099 makes SageTV work.

This removes other codes in the 14xxx range since they aren't the bruno
remote, nobody seems to know which remote they refer to, and they conflict
with the remote I *am* able to test and which may be useful to dogfooders.

Change-Id: I667af8db202b81f90dca8bbcb2ececf93c047492