S03usbinstall: bring back the install-from-USB-on-boot feature.
This got removed when we switched to simpleramfs.  Now it's back, although
(unfortunately) it requires both the kernel+rootfs now, not just the kernel.
If ginstall didn't require python, it could fit on the simpleramfs, which
would maybe be nice.  But a corrupt rootfs with working kernel seems to not
be a very common problem (I have not yet heard of anyone having that problem)
so this is a lot better than nothing.

Note that this operation should be safe to leave in place even for public
distribution, because ginstall and the bootloader should take care of
validating the actual installer images.

If it successfully installs an image, it stops permanently.  The old version
would reboot, but that doesn't seem wise; left unattended it would just
reinstall the same image over and over, and (without a serial console) you'd
never know when it was safe to power off the machine.

We also do a weak filename-based check of the install image version number.
If the file is named $(cat /etc/version).gi, we don't attempt to install it.
This is a secondary protection against repeatedly reinstalling the same
version.  It also means you can boot from USB, install the version on the
USB, then reboot again even without removing the USB stick first, and end up
running the software as expected.

If there are no images found worth installing, it unmounts the USB and
continues booting.  This is unlike the old version, which would stop
the boot process if there was ever a USB disk plugged in at boot time.

Change-Id: I2861ca2280de6bdd1bc1ca4dcc229551dcc9bcae