builder.py: with --no-build, set up the toolchain at least.
This builds 'worldsetup' but not 'world', which makes sure the toolchain is
ready to go and all source packages are checked out.  Then you can go to the
output directory and try to 'make -j package_name' for any package_name you
want to test to make sure its dependencies aren't missing anything.

Change-Id: Id045036711e270983cb13262d45fd149af30a092