Use gpio-mailbox to change the led colors.
Now that we have gpio-mailbox, let's change the led colours in a few places.
This will probably change around later, but it's a good test for now.

Specifically, we have a blinking blue light during boot, and a solid blue
once we get NTP sync, which is a decent proxy for "we have an Internet
connection."  We also blink the lights dramatically during S05usbinstall;
once it finishes, the lights either turn off (time to reboot) or blink red
(failed).

runminiclient changes the colours depending what program it's running.  This
is sort of pointless so it'll probably change later.  It turns the lights
off during waitpower though, which makes the device look off, even though it
actually isn't, which is I guess what we want.  Theoretically we could turn
on the "standby" light for this, but that's kind of obnoxious; that light is
really bright, and turning off your device shouldn't light up the room.

The *actual* spec for the lights is at http://wx/BrunoLeds and this is not a
complete implementation yet.

Change-Id: I33e593819bca6a4957ebe2979f88b99353510152