toybox/nice.c: don't bother with getpriority().
The nice() syscall returns either 0 for success (very old Linux kernels) or
the new niceness (which might be -1).  To check whether it was successful,
the nice(2) manpage says you should zero out errno, then if nice() returns
-1, check that errno is nonzero, and if it is, that's your error.

We could call getpriority(), but the way it was being done is incorrect.
getpriority() includes the *absolute* priority of the process, while nice()
adjusts the priority incrementally.  To use getpriority() correctly, we'd
have to get the priority before and after, and cap it at the
kernel-dependent range, and make sure the *delta* is the one we requested
with nice(2).  This is unnecessary since the return value of nice() is
perfectly sufficient.

Change-Id: If262f873a67b2e43e5333286d1ce3b08bae16594