stresstest: pkillwait dhclient before continuing.
dhclient can change the local IP address if the lease expires, thus deleting
our hardcoded 169.254.* address and interrupting the stresstest.

A "correct" fix might be to fix dhclient to not touch addresses that are
none of its business, but since stresstest isn't running on a "normal"
network anyway, this should do the job.

Change-Id: I7c238a32504c1dc842f0f4d1ed9ac6fada7049c1