coredump: get rid of "interesting" exes concept.
That hasn't done us any good.  The intention was that if multiple programs
coredump at once, we'd keep the "interesting" one (because they all go to
the same file and therefore would overwrite each other), but that turns out
to never happen.  On the other hand, it turns out that all core dumps are
kind of interesting, so we keep having to patch the script to add new ones.

Instead, let's prevent churn by only allowing one coredump every n seconds
(currently 600).  If you want one sooner, you can manually "pkill
coredump-delay".

This is actually better anyway, because the first process to coredump is
probably the interesting one, and we weren't actually preventing churn
anyhow if an "interesting" process was dying in a tight loop.

Change-Id: Ib3b559d839e8fe5f2d046b07b7936d1917ba79ee