upload-logs-now: a script to immediately end the upload-logs timeout.
Previously, we'd only upload logs 5 minutes after booting.  In case of
serious problems (like the system crashing a minute after boot, or users
that are annoyed by the blank screen and unplug/replug their box), this
might cause important crash data from the previous boot to be lost.

So let's allow arbitrary scripts to trigger upload-crash-log whenever they
want.  The first place to do it is in dhclient-script right after getting an
IP address, which is the earliest time we could possibly upload logs on a
new reboot.  We also do it after doing a core dump, in case the system is
about to fall down around us.

Change-Id: Ib46d528a120afaa257cb2c1f1685f0085de84b91