stresstest: allow overriding the server IP address(es).
Set $SERVER_IP on startup to just hardcode the address of the fat bruno
instead of auto-discovering it.  In theory, this remove the need for wiring
every pair of brunos on its own subnet for stress testing.  If SERVER_IP is
unset, we just broadcast to try to discover the server, as usual.

The new catawampus selftest extension also supports setting this.

Change-Id: I4e29d85990402a2e9d7ce12108ff39716aac21a4