Add partition version check in init.
As discussed with Avery, we will use hnvram variable PARTITION_VER for
partition upgrade. PARTITION_VER serves as a mark from the most recent
repartitioning operation.
If the PARTITION_VER is not set initially, the init script will set it
based on the current image's partition version. If PARTITION_VER does
not match the current image's partition version, the recovery code is
invoked. Each partition version has a specific handcrafted recovery
code.
Since the absence of PARTITION_VER implicates the latest partition map,
we need an intermediate image to initialize the PARTITION_VER. For the
first 4G to 1G migration, a two step approach is taken:

1. Upgrade with kernel change which recognize both v1 and v2 partitions,
and the init script which just initialize PARTITION_VER to 1 in hnvram.
2. Upgrade with init script with recovery procedure. This image will
download the same image of the current image and reinstall.

For the futural repartitioning, we will just need to combine 1 and 2
into one step. The above approach will cover both upgrade and downgrade
path except for all releases before the next release which includes this
CL.

This CL just address step 1 and after an official build, we will check
in the change for step 2.

In this CL, we split mounts to mount special purpose directories and
rootfs separately.

Change-Id: I95e6310e030bf6263d960a2e9a7a786eb0b230bb