dhclient-script: ignore EXPIRE events, fixing HDHomerun bootup problems.
Here's what would happen:

- We get a DHCP lease
- NTP syncs and causes a time jump forward
- Other processes (like chsrv) see that NTP is okay, and start working
- Our DHCP lease expires instantly
- dhclient-script runs and brings down the interface
- dhclient gets a new DHCP lease a few seconds later.

This caused a network connectivity glitch for somewhere between a few
milliseconds to a few seconds, depending on the speed of the DHCP server.
During that downtime, sage, sagesrv, and chsrv would all start, but, if you
were unlucky, would not be able to connect to the network.  Under some
circumstances, chsrv would get confused about the existence of your
HDHomerun device and register the HDHR-CC-*-* channels as HDHR-NR-*-*
instead, and then drop them because it only supports CC channels.
Unfortunately, it doesn't exit in that case, even though it has no encoders
left.

For some people, this happens more than others (I think it has to do with
how slow your local DHCP server is).  Rebooting your fat bruno
repeatedly would change the timing enough that it wouldn't happen anymore.

A different, but similar problem would occur if the lease was dropped before
chsrv got any answer at all.  In that case, it wouldn't even see the
HDHomerun device, so it would exit after a few seconds, and babysit would
restart it after 60 seconds.

Change-Id: Ibfee151748bcdefe331e4bb3bc4f514b5357f3c8