Fixed bug + reordered checks.
The ipv6 address should be checked only when ip.changed changed.
If the file causing watch-dirs to trigger is dhcpbound and there was an
ip address, echo-ing into dhcpbound would add a new event to the queue
and would make watch-dirs trigger again. It will go through the same
path, so there will be an infinite loop, causing inotify queue overflow.
Also, changed the order of the file checks to allow dhcpbound and
acsconnected to co-exist and show the correct sequence (which should be
ACS_LED_ON).

Change-Id: Ib0b4ce38998020effe9a0782eb772abcb14cdce3