Add poweroff and poweroff-with-message commands.
poweroff-with-message is tricky: we want to print a message to the log, then
make sure the log has been uploaded to the server (if at all possible) since
it will be lost by the poweroff.  That required some rearranging of the
upload-logs stuff.

While I was there, I modernized S95uploadlog a bit, so that its stop/restart
commands actually work correctly.

To implement the feature, I added the new commands leds-and-wait (like
'leds', but guarantees gpio-mailbox has updated the leds before it returns)
and upload-logs-and-wait (like upload-logs-now, but waits for the upload to
complete before it returns).  leds-and-wait requires a slight change to
gpio-mailbox, which is a separate commit in the platform package.

Change-Id: Icb6519c67cd3e0d152696051987431496329204c