loglinear: another bugfix, and a test.
It turns out the logic in loglinear is kind of complicated! :)  Found
another condition in which it would fail to print output for an errored
subprocess.  I'm getting a little tired of this, so I added loglinear_test,
a script that checks to make sure it works as expected, which it now does.

Change-Id: I93a87f80c6013b9e7036b958a9e674e457e54db6