Reset BT chip before starting bsa_server.
If the BT chip gets into a bad state, bsa_server cannot run. A reboot
or soft reset does not reset the BT chip. Currently, the only recovery
mechanism is to power cycle the TV box. This CL is to explicitly reset
the BT chip before starting the bsa_server to ensure that the BT chip
is in a known good state.

In addition, this CL combines S45app_gtv and S30bsa as app_gtv has
explicit dependency on BT chip being successfully initialized by the
bsa_server.

Change-Id: I10a9b87bdc8d85ebb0d8264384efb6319ee4f504