simpleramfs: make init=/bin/sh kernel option work again.
mounts-root was relying on a file named /tmp/initcmd, but /tmp isn't mounted
in the simpleramfs (it mounts in /rootfs/tmp).  Rearrange things a bit so we
don't need to rely on /tmp.  This makes the code a bit cleaner anyway.

Also, when things go wrong, reboot immediately rather than giving a shell.
You can still get a shell if you control the bootloader, by setting
rdinit=/bin/sh.  But if there's an error, we don't want to accidentally leak
a shell to an end user of a supposedly locked device.  Also, for 99% of
people, failure to mount the rootfs is not something they'd use a shell to
fix; the correct thing is to reboot and let CFE go back to the alternate
software image.

Change-Id: I1e0b42f8ac0b0cc567da50ff9016d795dba45b6b