Disable login=1 by default. Its time.
By setting login=1 in DEFAULT_KERNEL_OPT, ssh and serial console login
access is enabled... except that we've had it enabled by default since the
beginning of the dogfood program.

As we're now approaching the release, we need to lock down the firmware.
In particular, we're about to cut a new manufacturing software image.
Login should be disabled for freshly manufactured units.

Change-Id: I610035da51404c1caabff3de2ea22f1e572b5a22