Switch the mounting order of config and user.
If the number of bad blocks increases, the leftover bad blocks will
not be sufficient enough to mount config partition. The solution to
this is to guarantee config partition's quota before mounting /user.
The current partition should not be changed, only when the config
partition mounting fails, this script will kick in and reverse the
order of volumes. Since most of our code look for volumes based on
name instead of ubi/mtd number, we should be fine.

Change-Id: Iadeea61b1ac932a7adab4ac2b611970985533234