configs/bruno_gfhd100b2_defconfig: rewrite using 'make savedefconfig'
I didn't know about this feature before, but it turns out to be exactly what
we want:
	./scripts/builder.py -x --no-build
	cd ../out
	make savedefconfig
	cp defconfig ../buildroot/configs/bruno_gfhd100b2_defconfig

'make savedefconfig' writes only the .config lines that differ from their
defaults, so the file contains as few options as possible and is much easier
to read and should have far fewer merge conflicts.  When scripts/builder.py
uses that as the defconfig, it still produces the same .config as before.

Change-Id: If09f415f5109e85a5ea1e9f1c1b543c7b6e8061a