samba: run under babysitter, and actually log stuff
I don't know where the logs were going exactly, but they weren't going to
the log (maybe the default loglevel was 0).  Now we also restart it in case
it dies, and add a couple of minor settings that could slightly improve
security.

It still leaves force user = root, which is a serious security problem on
its own, but will take a bit of testing if we fix it (I think it'll interact
badly with sagetv, based on trying it briefly before).  I would also like to
set 'root dir = /var/media' but that'll take a bit of effort to make work
well.

Also disable samba's locking feature, which generally makes it slightly
faster and avoids complication.  It's only necessary for things like
Access databases, which I doubt anybody will use our device for.

Change-Id: I1e968a044639cb2964d475d58eb1fe8343da804b