mkmakefile: don't run loglinear from out/Makefile
That was kind of useless: if you do
	cd ../out
	make something
then the all: target would call loglinear just to jump back into
../buildroot/Makefile which already calls loglinear for its submakes.
That was especially bad because things like 'make oldconfig' would then not
work right.

As a bonus, this avoids the ugly buildroot+all level in the
loglinear hierarchy, if you were building from the out/ dir.

Change-Id: I2df3f2a9371f6e4e063ce72fb3db261ca7e13e29