dhclient: set hostname to GFiberStorage vs. GFiberTV.
Previously, we just always set the dhcp client hostname to GFiberTV, even on
the storage box, resulting in incorrect listings in DHCP server status
screens.  (There wasn't much other impact, but it still makes things hard to
figure out.)

For the primary device hostname, renamed gfiberstorage->GFiberStorage and
gfibertv->GFiberTV.  It's less unixy that way, but looks better in DHCP
server status screens, which is how I determined that everyone else does it
that way nowadays too.

Also removed the explicit netbios name from smb.conf; the default (use the
local hostname) should always be what we want, assuming the local hostname
is scrubbed properly.  And changed the samba "server string" to be a longer
text string since it's intended as a description, not a hostname.

Change-Id: If3470a9374799f40ed124a6b71fbdaa8d864c36a