add null download method.
The download method for simpleramfs is "/dev", which falls to the
default download process logic which invokes wget. The 1.12 wget is kind
of buggy and returns 0 even if download scheme is missing. For 1.13 wget
it will return 1. We don't have this issue if we use ubuntu 11 for
ubuntu 12.04, the build will fail at simpleramfs.

Change-Id: Ie5023aa7f428386685f88d9d1f61bd77bd2e95cd