Add tune option for marvell chip.
- Get rid of the following warning.

"This Marvell-tuned GCC toolchain may not be sensible for non-Marvell
cores. "

- Remove the configs which automatically come with ARCH.

- We basically need to patch the Config.in.arch to give a dummy prompt for
GCC TUNE. Otherwise it will always pick up the default value. This is from
kconfig document:

"The default value is only assigned to the config symbol if no other
value was set by the user (via the input prompt above). If an input
prompt is visible the default value is presented to the user and
can be overridden by him."

Change-Id: I899bcb34dfb85abe633ef24ae8ec80e06a39b8f6