S91smb: chown /var/media/$d, not just its subdirs.
The previous S91smb patch was almost right, but when we recursively chowned
a directory, we failed to chown that very directory, and only did its
children.  Symptom: you could not create files or directories in the root of
Pictures, Videos, or Music.

The *actual* bug is that my fake implementation of find (since we lack a
real one on bruno) fails to print the base filename in addition to the
subdirs, unlike "real" find.  (That is, the first line of 'find .' should be
'.' and it isn't.)  I'll fix that too, but changing our implementation of
find could have unforeseen consequences, so I'll do this first as it's
suitable for backporting.

Change-Id: I0f61d7a4e40af8f1ca70592f7c3f1c58796ac3a0