Make current time survive warm reboot b/7374977
The whole point of this CL is to allow the boxes to boot normally even
if they don't have NTP access.
For that I added a new flag to logmark-once indicating whether NTP is
connected or not. Immediatelly after reboot, it looks for the last last log
message that had the NTP flag set to 1, and set the current time based
on the timestamp of that log message.
Also, I changed the sageserver and startupvideo initialization scripts so
that they don't wait for NTP to be set before starting, it's enough
if the approximate time mentioned before is set.
Change-Id: I79c9a4d83c64e44083ab7075cc79fcdeecaf159d