openntpd: set minimum time skew for settime() to 10 secs, not 180.
The recent change to set the system time at startup based on time recorded
in the PRINTK_PERSIST log had an unintended side effect: the time at boot is
now *close* to right (within about 40 seconds), but not quite right.  That
meant ntpd would skip the settime() step and just adjtime() instead, which
meant it could take many minutes after booting before the clock would
actually be correct.

Being a minute off could be seriously annoying for a DVR.  Even the 10
seconds is kind of a pain, but it should fix itself within a few minutes
(fixing 10 seconds of skew is much faster than 40 seconds).  Moreover, since
the time from PRINTK_PERSIST is *more* than 10 seconds off, this makes
settime() actually happen, so in practice it will be correct within
milliseconds, as soon as the ntp server connects.

Change-Id: Iffe621e4fc5539592bbf2ca1f389f1deced6df9c