softdog: avoid mem leak, reap finished backgr jobs
This commit is related to b/7924361 but does not fix it.

softdog creates background jobs using an ampersand (&) at the end of the
line like in

hnvram -r PLATFORM_NAME >/dev/null &

dash saves an instance of "struct job" for every such job. It only
releases this struct under certain conditions like when the "wait"
command or the "jobs" command is issued.

softdog creates background jobs in a loop without reaping finished jobs.
As a result, the job table grows indefinitely. On a thin bruno that had
been up for 6 days, a heap size of 1432 kB has been observed. This is a
drastic increase from 16 kB which is the size of the heap right after
softdog starts.

We solve this problem by issuing the "wait" command unconditionally and
with every iteration of the loop. This does not change the logic of the
program because there should not be any running jobs at this point
anyway.

Change-Id: I111aead2fd256d5a4edd685f5e41ec39f6b049fd