If the server 503s the request, then retry.
Retry the upload with an exponential backoff,
this is to try to spread the load during the
"stampeding herd" when a large swath of machines
all come back to life at the same time.

upload-logs-now will first kill upload-logs
will interrupt the sleep() call and cause upload-logs
to run right now, which will upload the most recent
log data.

Also removes the retry from the higher level upload-logs-loop
script.

Change-Id: Ibf57793d37a006fc75ff22c2e786d69bf227c617