upload-crash-log: also add a logmark after starting the upload.
After a reboot (intentionally or due to a panic), the last log segment often
didn't contain a logmark (which contains, most importantly, the software
version).  So when we rescued the log chunk after the next reboot - which
might be into a different software version - the log upload server didn't
know for sure which version that segment came from, and would assume an
empty string instead.

The only reason any panics were ever associated with the right version is
that logmark runs every 30 seconds anyway.

Beware, this patch will make it look like the kernel is panicking a lot more
often in versions that include it :)

Change-Id: Idf5608af34435732547b9b70b0d013d4c5ed85bd