/bin/coredump: refuse to make a coredump file for siege.
This is a bit heavy-handed, but right now we have a lot of systems that end
up stuck for 120+ seconds trying to coredump the siege process, which is
suspicious (and might be related to running out of memory).  That's not
helping.

I checked and I didn't find any other processes that caused
the kernel to trigger its 120-second stuck task panic when they dump core,
so those should be ok.

Change-Id: I540560d37d0fb29fa9b17de2ea4d80090e47c6e7