Add procps patch to fix unknown HZ value problem.
Sometimes, we will see "ps: Unknown HZ value! (73) Assume 100.". (73)
could be any other value. The root cause seems to be related to gmake
which breaks the init sequence during procps bringup. Two issues are
caused by this problem as described below. The first one is addressed by
procps-make-3.82.patch so that the build went correctly. This CL is to
address the second issue.

http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=603759

GNU Make 3.82 no longer sorts lists of filenames when expanding
automatic variables or wildcards. Filenames are now returned in
filesystem order. This new behaviour has broken the procps Makefile.
There are two problems:

1) In proc/Makefile at line 177:

-include */module.mk

When that wildcard is expanded as "ps/module.mk proc/module.mk", the
makefile fails to link ps/ps (because LIBPROC is empty).

2) In proc/sysinfo.c and proc/version.c:

The functions init_libproc() and init_Linux_version() are both
declared with __attribute__((constructor)). So they are automatically
called before main(). And since they are not prioritized and will be
in separate object files, their order of execution is determined by
the order in which their object files are linked. Which is now less
predictable.

init_libproc() uses the value of the variable linux_version_code which
is initialized by init_Linux_version(). When init_libproc() is called
before init_Linux_version(), hackery for legacy and non-Linux systems
is called upon (old_Hertz_hack()) because linux_version_code is still
zero (and <= LINUX_VERSION(2, 4, 0)).

And then we get messages like:

Unknown HZ value! (67) Assume 100.

Change-Id: Iea8da2855092661166f62e24a4d080c1ddccc84a