Slightly improve boot scripts for booting on other platforms.
fs/simpleramfs/mounts-root: remove some ugly unnecessary braces, and allow
root=/dev/whatever direct pointers to a device.  If the root= does not start
with /dev, we keep doing what we always did, which is try to mount an ubifs
volume.

fs/skeleton/etc/passwd: we previously had a variant of this per platform.
There's no particularly good reason to do that since there are only a couple
of different users.  Just make one file that's the superset of both, and we
can conveniently have a user list that's identical across all devices.

Change-Id: I5e3933d560519c9c4b2fe82231bcfd3fe166375b