Merge BR2_PACKAGE_BRUNO into BR2_PACKAGE_GOOGLE_PLATFORM.
Most of the stuff in the "bruno" package (eg. logos) is not really
bruno-specific and can be used on any of our platforms.  And the
google_platform was actually just an empty placeholder that didn't do
anything.  So move it in there.  As a bonus, now the package name
(google_platform) matches the repository name (vendor/google/platform).

While we're here, overhaul the way we differentiate between platforms.
Instead of having a single item that controls both hardware configuration
and software features, split them apart.  Now you can choose separately
whether you want this software image to work as a fiber jack, network box,
storage box, or TV box (or any combination of the above).  Not all
combinations actually work correctly yet, but at least now there's a
framework for getting there eventually.

Furthermore, we create some metapackages that depend on other standard
packages.  This lets us remove some options from the defconfig files, since
they are the defaults if you enable the relevant metapackage.

Change-Id: I40e25e789e1df7f4880f50395afc0b0d978fca21