kvm_defconfig: enable google_hnvram package.
With an emulated mtdblock device, we can actually use unmodified hnvram
inside a kvm session.

With hnvram in place, we can set a fake serial number and model number.
If we pretend to be a GFMS100, that seems to be enough to get sagesrv
launched.  There's no jvm yet so SageTV itself can't load.

Change-Id: Ibf52f9897f10db15ab179fd53dd58e76f47e31a8