Move Makefile logic from buildroot to repo
Every time we want to change how software from vendor/mindspeed/drivers
is built, we have to change the corresponding package in the buildroot/
repository, too. This is tedious because it involves changes to two
different repositories. Let's try to keep the .mk file in buildroot/
static and move Makefile logic closer to the actual source files.

Change-Id: I75f88999c879379ec96a139b34051db7d10f7f88