Teach fs/ginstall how to build ginstall images containing a uImage.
We want the GFRG200 to be able to accept ginstall images and eventually
share the same ginstall implementation with GF{HD,MS}100.  Unfortunately,
the kernel format for CFE and uboot are quite different (vmlinux.gz vs.
uImage), so we need to support both in *.gi files.

One nice thing about uImage files is they can have a separate initrd file,
rather than linking the initramfs into the kernel binary as a .o.  That's
handy because it gets rid of the annoying circular dependency between
building the kernel, and building modules that depend on the kernel but
which might want to be in the initramfs.

Change-Id: I8c612acb59791862ede9cb815dad7540c44e586d