Fix autobuilder error in kvm_defconfig caused by previous patch.
Now that the skelmir package is selective about which arch binaries it
installed, it drew attention to the fact that we were accidentally
installing the MIPS skelmir VM on x86_64, where actually openjdk (which was
also being installed) should be sufficient.

So let's make google_sageserver depend on one or the other, but not both.
It's also a config depend, but not a build-time depend, because the jvm
isn't actually needed to build sageserver, only to run it later.

Change-Id: I0cb718b781900a95892120d1ed0af397a7c6ea0e