Bluetooth firmware update.
bsa_config.hcd isn't really a config, its microcode for the Bluetooth
chip. Rename it to BCM20702, the name of the chip.

Add firmware for the BCM20705.hcd, used in GFHD200. The two chips
are not compatible, they need separate firmware.

Use the correct firmware in /etc/init.d/S30bsa.

Change-Id: I008761cc86d8f5db47c54acb22de2d67ef705dc3