Added onu-specific ledmonitor.
LEDS are controlled by the leds-gpio brightness and timer trigger interfaces.

LEDs assume one the following states: off, solid, blink (1000ms on, 500ms off),
and fast (500ms on, 500ms off) and only one LED is on at a time.

The monitor stores the current state off the LED to minimize logging, but it is
otherwise stateless, so if the LED state is 'solid red' and something changes in
the monitored directory but the resulting LED state is still 'solid RED', the
monitor will still call the kernel interfaces to set the LED state, but in a way
to avoid glitching.

To handle the case where the ledmonitor is restarted and something has changed
in the monitored directory in the interim, the while loop tests whether the
led_state variable is null so the first pass always evaluates to true.

Change-Id: I2587d60754c01c728d9f294869edeaf77d128d67