Samba fixes.
Syslogd is no more. Log to stdout.

Load the vfs extd_audit module for more logging of
activity, to better diagnose problems.

chroot samba to directories in flash, not the drive.
The drive can be removed from the device and modified,
and samba is now going to be loading compiled modules
(like extd_audit) from it. This would be a security
problem.

Set "unix extensions = no" to fix MacOS file permission
errors b/10530244

Change-Id: I66a79717008d599bc1a8c2bc950ebbc1620b1c2d