Add libcurl and openssl as dependencies to sagesrv
These dependencies are required by the ad loader service that uses
libcurl for http download requests and openssl for computing the sha1.

Although right now all the ads code is in a separate branch
(remotes/gfiber-internal/ads), I think it doesn't make any sense to
create a separate branch just for this makefile update.

This change doesn't harm in anyway the sagesrv code or functionality
as it is right now.

Change-Id: I704fcb4737badc2b083c5b29ed681b3ccbeebcde