mindspeed_drivers: Change Makefile for FCI and CMM
Change this packet to an autotools package because FCI and CMM are
autotools based. The mindspeed_drivers repo now contains a fake
./configure script that recurses into fci/lib and cmm.

We had to fold the install step into the build step because fci/lib
needs to be installed in the staging directory before cmm can be built.
We defined this dependency in the Makefile in vendor/mindspeed/drivers.

Change-Id: I4d8d4d4d0c6488cdee86265f40a198468866fe18