Fix fs/ginstall and copy-skeleton to work with recent changes.
The builds were succeeding, but output was confused. When PLATFORM_SUFFIX
was repurposed, the code for auto-renaming files in fs/skeleton that had the
.platform_whatever suffixes got totally broken.  This fixes it up for
GFRG200, GFMS/HD100, and GFHD200.  I haven't tested GFLT though.

Also, ginstall.mk was creating v3 images that were hardcoded to support the
GFHD200 platform, which was quite a surprise when I tried to load one on
GFMS200 and it worked.  Fix that up by adding an extra field in the config
file.

Also, ginstall files on GFRG200 were messed up because of confusion between
kernel input and output filenames.  Clear that up too.  I'm not sure the new
files on GFRG200 actually load (and ginstall is still broken on that
platform) but at least they look more or less correct.

Change-Id: I0a85bc9ef307b163f6b7012a848157a3d7a5fa64