simpleramfs/mounts-root: don't be overly specific about rootfs device types.
It's not really necessary to check the specific type of the rootfs0/rootfs1
devices.  In particular, checking this introduced a bug that caused GFHD100
devices to take longer to boot since mtd devices are character, not block,
so the 'wait for device to appear' loop had time time out before continuing.

Change-Id: I4f1bf2fc4c167428616eb6c7dbeb1ed01fd728ad