copy-skeleton.sh: handle a few more platform prefixes.
Autobuilder revealed that the last patch fixes gflt110, but disallowed
gftv100 (which uses a gfibertv prefix for legacy reasons) and gfrg200.  It
also uncovered a bug where we weren't setting a platform_prefix at all for
kvm builds.  And postbuild.sh wasn't aborting when copy-skeleton.sh failed,
so builds would sometimes be aborted for some other secondary reason
instead, so add a "set -e" to catch it sooner.

Change-Id: Ibf3929e471d236117c84974dac546948d304c383