wifi: refactor fccwifi to support both Atheros and Broadcom chipsets.
Also rename fccwifi to just 'wifi'.  Eventually I want to call this from
catawampus rather than teaching catawampus about both (or many) kinds of
wifi chipsets.  And while fccwifi was used for FCC certification, it was
always ugly that code was duplicated between catawampus and that script.  It
still is, but this gets us one step closer to merging them together.

Change-Id: Ibd2922e32a55341d47e9252fb7188225e6421b4e