Cleanup of if{up,down}.br0 and babysit; add startpid and killpid.
if{up,down}.br0: automatically skip running dhclient if the interface
already has an address.  (This happens when using NFS root.)  That lets us
simplify the code in S40network so it can just always run 'ifup br0'.  That,
in turn, is important with NFS root nowadays, because we wouldn't properly
set the interface address on br0 when booted with NFS root on GFRG200, which
didn't really make sense anymore since NFS was on wan0 anyway, not br0.

While we're here, let's move dnsmasq from S40network to if{up,down}.br0 so
we can restart it on demand.  And while doing that, let's clean up the
ifup/down scripts.

Also added two new commands, startpid and killpid.  startpid starts a given
program and writes a pidfile containing its pid, for later killing with
killpid.  This is safer than using pkill all the time.  In particular, it's
important if we need to be able to start/stop particular instances of
programs (like hostapd, dhclient, or dnsmasq) bound to particular
interfaces.

Modified 'babysit' to know how to skip over certain "prefix programs" when
printing error messages.  Previously, 'babysit 10 startpid /tmp/file false'
would give messages like "babysit: 'startpid' exited with error code 1"
which is technically true but not useful.  Now it knows how to skip over
startpid, alivemonitor, and other babysitter instances so it can identify
the program as 'false' instead.

Change-Id: I180147487f4bf0b82550cb2a1adda994a8278e38