gfrg200: officially switch from uboot to barebox.
We've actually been using barebox for a long time, but it wasn't being built
as part of the main buildroot.  uboot doesn't even boot on this device
without changes, so disable it.

Also add a new barebox_uloader package that builds from the same barebox
source code, but with a different defconfig.

Change-Id: I56a31c0453b7fb2351d470e6c239f0c27f99e9b6