S06usbinstall: reject *.gi files for the wrong platform.
The old behaviour was mostly harmless, in that ginstall already knows how to
reject files for the wrong platform. But if you had two install packages on
the same USB stick, one for each platform, this script might pick the wrong
one, then ginstall would reject it.  Now it'll skip the wrong one, allowing
it to pick the right one instead.

Change-Id: Iea839fba071356ff085c50de5c8b8e734fac6a22