Disable apm warning messages:
In the gflt110 logs there are bunch of lines like:
apm_avc_main_loop: type 7, param1 0, get state Failed!

Per the Marvell engineer:

AVC type 7 is for IP address, the printing indicates it fails to get the IP address (the interface is fixed to be “pon0” now).
The below is the implementation of this callback function, it is possible that Google does not have “pon0” interface?
Anyway could disable these AVC functions by XML
/etc/xml_params/apm_xml_cfg_file
        <!--  AVC enabled/disabled flag, true:enabled AVC functions, false:disabled AVC functions. -->
        <apm_avc_activation enabled="false" />

Since we don't set an IP address on pon0, the function fails and it spits out that
warning.  This change to the xml is supposed to disable AVC and stop this error.
I have a question open with Marvell what AVC is what else it does.

Change-Id: Ied264e2d38c2e346e527f8702e567240fdd7bb50