If PRODUCTION_UNIT=0 leave telnet enabled.
If PRODUCTION_UNIT=0 then do the following;
- blink both red and blue leds.
- Don't start the prodticker count down.
- Enable telnet.

This is a special manufacture mode that Actiontec requested to
allow setting a unit in this mode and leaving telnet enabled for
the duration through multiple reboots.

Also as a bonus, change setprodflag script to kill off inetd and
telnetd after the timer expires.

Change-Id: I46df2ec663f578cd525f4a15c263163cfe387507