Makefile: pass $(EXTRAMAKEARGS) correctly in a couple more places.
Without this, O= was not being set, so the sub-make was using the default
buildroot config for i686 instead of the actual $O/.config file.  Mostly
this was harmless (because the submake wasn't doing much with the config)
but it caused a symptom where google_dart_vm.mk would raise an error about
the wrong architecture.

Also rename the BTYPE variable in google_dart_vm.mk to prevent potential name
conflicts with other makefiles.  All make variables are global in buildroot.

Change-Id: I5e3625ad11ea39e627e941060afef3e4a22f891c