HACK: change US 5 GHz power limits to 0 to appease ath10k driver.
Without this, transfer speeds are severely reduced now that we're loading
the actual correct power limits, which greatly improves the ath9k but hurts
ath10k.

The RSSI received by my laptop with the zero limits is -48 dBm.  With the
"real" limits it's around -34 dBm.  The higher power level seems like it
should be better, but maybe it's causing distortion.

Change-Id: Ieb59b33cd8acf313fe342c6b5b673219cdb69443