S94startupmods: split this out of S95sageserver.
It was annoying that restarting sageserver would re-run startupmods over and
over, which is almost never what you want.  (It would generally be used to
mount a filesystem or start a daemon, and it would get started over and over
again.)  It also resulted in duplicated code between S95sageserver and
S99oregano.  Also, if sageserver and oregano weren't enabled in a build,
startupmods wouldn't work either, which eg. stops them from being usable in
opensource builds.

The new script runs /rw/startupmods and then /rw/poststartupmods, one after
another, before sageserver or oregano actually start.  This sounds wrong,
but it's what was actually happening before, because the segment of code
between them was run in a subshell in the background, so we didn't wait
until it finished before running poststartupmods.  Thus there was a race
condition where both sageserver and poststartupmods were starting at the
same time.  Now we wait until poststartupmods is done before starting
sageserver, which is one of the things that could have happened during the
race condition before.  (It makes poststartupmods basically useless, but
it's probably not a good idea to change its behaviour now.)

Change-Id: If36be10a5f4503f897d5bd5b8d1cdf2edd5e47a5