Whether or not you need hnvram is a hardware issue, not a feature issue.
You don't need hnvram because you're a TV box, you need it because a
particular device uses it for its settings.  The way we were doing it,
building a platform-only (builder.py -x) build would not enable hnvram
because a platform-only build is not a TV box, storage box, etc.

Change-Id: I20eafdba8c1417a62007efd4752d83551ec9c073