Add a vlan2 interface for the wan and start dhclient.
For the wan interface create a vlan2 tagged interface,
bring the interface up, enable nat to that interface
and start dhclient on that interface.  This should all just
work correctly as long as only one of the two interfaces gets
a valid dhcp response, and there is one valid default route.

Change-Id: Ia65ebd57beac2b59e906a373683115a7a9fdf29b