mwifiex: for google fiber create just the AP interface, call it wlan%d
To work around a bug in mwifiex which doesn't allow changing the type
of the interface (which causes hostapd to fail), create just one AP
type interface at mwifiex driver initialization or card insertion.
Also, call it wlan%d so that it is not gratuitously different than
other cfg80211 wireless interfaces (though being of type AP is a bit
different).

Change-Id: I80f179249d6219014c201f9deeff901ea6d888f6