boot scripts: slightly smarter waiting for scsi devices.
If the system booted really quickly, S06usbinstall would sometimes finish
before a usb-storage device even finished probing in the background, so it
wouldn't be able to install.  S60harddisk could have been susceptible to a
similar problem.

Add a new script, wait-scsi, that checks all the available scsi interfaces
on the system and waits a few seconds if any of them don't have a disk
attached.  Then, to minimize boot latency, we run it in three places:

- S06usbinstall: only waits for USB interfaces
- S08volcheck: starts polling all interfaces, but in the background
- S60harddisk: wait for any interfaces that still haven't timed out from
   the polls startedin S08volcheck.

Also, don't mount hard disks unless the device is intended to be a storage
box.  Fixes b/12688028.

Change-Id: I9fe069d26b9d2a24d8bf3fb73c7c39639fb72056