ath10k: HACK: force txpower level to 0 dBm.
This isn't *really* 0 dBm, it's just the value that causes our wifi module
to transmit at the right level.  If we allow the driver to adjust the power
level, it ends up oversaturating and causing distortion and poor signal
quality.  This isn't a real fix, but it works for now.

Previously we implemented this by setting the regulatory limit to 0 dBm, but
that inteferes with non-ath10k drivers such as mwifiex.

Change-Id: Ic58deb60e76a40d53143f151cf4a586967ef7c16