/bin/wifi: reduce one source of ath10k crashes.
The ath10k driver often crashes when we kill hostapd.  Obviously we need to
fix that, but in the meantime, it turns out that if we down the interface
first, it makes this particular crash go away.  This doesn't cover the case
where hostapd crashes (which happens sometimes, sigh) but it helps a lot at
least.

Change-Id: I2443d115392c39b03d40f3c9b3356d405db2acc1