ath10k: a patch to improve stability, plus my patch.
After a lot of experimentation, it appears that the
warmboot-instead-of-coldboot patch by Michal Kazior does significantly
improve stability in our setup, particularly if we're using the ath10k 'ap'
firmware instead of the station-mode firmware (because the latter is very
crashy if you try to use monitor mode).  It's still not great, because cold
resetting the chip should be something we should feel safe in doing, but it
may be the best we can do until/unless QCA comes up with a better fix.

I also added a few extra log messages that may help diagnose any remaining
crashes.  It also reveals that the driver resets the wifi chip (and then
reloads the firmware!) way too many times during startup, which is why it's
so slow.  That should be pretty easy to fix, but I was hesitant to do so
since it would make the firmware crashiness less reproducible (since it
crashes more when you reset more).

Change-Id: I91649e363a5b0137d6ad596477d21432ed131909