Fix S06usbinstall to handle new Sandisk flash drives
New USB sticks from Sandisk and potentially other vendors are not
being reported as 'removable' media to the block device driver, so the
the S06usbinstall scripts doesn't recognize them. In the kernel logs
it says "Attached SCSI disk" instead of the expected "Attached SCSI
removable disk". When plugging one of these drives into a Goobuntu
workstation it actually registers itself as a "Hard Disk" and not a
"USB Flash Disk". After some research, it looks that Sandisk is doing
this too all its USB sticks now in order to comply with Windows 8
certification (http://goo.gl/ZnJmc1).

The fix is to not check for the "removable" flag anymore but instead
check if the device path contains a USB bus location.

Change-Id: I21f061e75d657f8f95aff191f32ba591f3fd166d